Use correct trigger data type. Add conversion for MicroDAQ.
MicroDAQ expects an int and PeriodicTrigger delivers an uint. The
conversion module does the trigger conversion in order to use the
PeriodicTrigger. This should be removed once the MicroDAQ module uses
uint too!

+++ include/LoggingModule.h
@@ -27,7 +27,7 @@ namespace ctk = ChimeraTK;
 struct LogFileModule: public ctk::ApplicationModule {
   using ctk::ApplicationModule::ApplicationModule;
 
-  ctk::ScalarPushInput<size_t> trigger { this, "trigger", "",
+  ctk::ScalarPushInput<uint64_t> trigger { this, "trigger", "",
     "Trigger used to update the watchdog" };
 
   ctk::ScalarPollInput<std::string> logFile { this, "logFile", "",
